bank-command-service/src/main/java/eu/whitebox/bank/controller/ExceptionHandlerController.java
package eu.whitebox.bank.controller;

import eu.whitebox.bank.exception.OverdraftLimitExceededException;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

@ControllerAdvice
public class ExceptionHandlerController {

    @ExceptionHandler(IllegalArgumentException.class)
    public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException e) {
        return ResponseEntity
            .status(HttpStatus.BAD_REQUEST)
            .body(e.getMessage());
    }

    @ExceptionHandler(OverdraftLimitExceededException.class)
    public ResponseEntity<String> handleOverdraftLimitExceededException(OverdraftLimitExceededException e) {
        return ResponseEntity
            .status(HttpStatus.UNPROCESSABLE_ENTITY)
            .body(e.getMessage());
    }
}